Refactor/login device id to object

+++ src/main/java/com/Ali/Store/App/dto/product/response/ProductDtoForCartItems.java
@@ -1,4 +1,6 @@
 package com.Ali.Store.App.dto.product.response;
 
-public record ProductDtoForCartItems(Long id, String name, int price, Boolean status) {
+import java.math.BigDecimal;
+
+public record ProductDtoForCartItems(Long id, String name, BigDecimal price, Boolean status) {
 }
